chore(dal): move kubernetes builtin schemas to their own module
