+++ lib/dal/src/builtins/schema/kubernetes/kubernetes_spec.rs
